+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 * Added support for nested controllers _(PR #422, Issue #283)_
 * Fix issue where the `failedValidation` property was not set for `Content-Type` request validation errors _(PR #420)_
+* Fix issue where `array` `body` parameters would coerce non-array types into an array _(Issue #438)
 
 ### 0.10.1 (2016-02-24)